Extracted validation to its own file.
